Documentation
At least, the first few steps of documentation. Still a lot more to do
at the moment, only header files are documented, I plan to document the
implementation differently, as a more of "this is how this does it"
rather than manual style. Also, there need to be a lot of refinements,
to my code, wow.