offset into kernel objects.
This is a property set on the constructor of kernel objects. As the name
suggests, any address generated for x will call the generate function
with x - offset. It will not be called whith values below 0.

The next step is to have shift_right and shift_left functions to adjust
the offset of an existing kernel object.